+++ apps/sandbox/src/app/async-table.queries.ts
@@ -72,6 +72,7 @@ export interface DiscoverMovieQuery {
 
 export const discoverMovies = client.get({
   route: '/discover/movie',
+  reportProgress: true,
   types: {
     args: def<DiscoverMovieQuery>(),
     response: def<Paginated<Movie>>(),
@@ -87,6 +88,7 @@ export const testCall = client.get({
 
 export const uploadFile = client.post({
   route: '/upload',
+  reportProgress: true,
   types: {
     args: def<{ body: FormData }>(),
     response: def<{ id: string }>(),