+++ libs/query/src/lib/request/request.types.ts
@@ -16,12 +16,10 @@ export type Method =
   | 'CUSTOM';
 
 export interface RequestError<Detail = unknown> {
-  code: number;
-  message: string;
-  detail: Detail;
-  raw: unknown;
-  route: string;
-  requestInit?: RequestInit;
+  url: string;
+  status: Detail;
+  statusText: string;
+  detail: unknown;
 }
 
 export type ParamPrimitive = string | number | boolean | null | undefined;
@@ -38,6 +36,62 @@ export type PathParams = Record<string, string | number>;
 
 export type UnfilteredParamArray = Array<UnfilteredParamPrimitive>;
 
-export type CacheAdapterFn = (headers: Headers) => number | null;
+export type CacheAdapterFn = (headers: RequestHeaders) => number | null;
 
 export type RequestHeaders = Record<string, string>;
+
+export interface PartialXhrState {
+  headers: RequestHeaders;
+  status: number;
+  statusText: string;
+  url: string;
+}
+
+export interface RequestConfig {
+  method: Method;
+  urlWithParams: string;
+  body?: unknown;
+  reportProgress?: boolean;
+  withCredentials?: boolean;
+  responseType?: 'arraybuffer' | 'blob' | 'json' | 'text';
+  headers?: RequestHeaders;
+  cacheAdapter?: CacheAdapterFn;
+}
+
+export type RequestProgress = {
+  current: number;
+  percentage?: number;
+  total?: number;
+};
+
+export type RequestEvent<Response = unknown> =
+  | {
+      type: 'start';
+      headers: RequestHeaders;
+    }
+  | {
+      type: 'download-progress';
+      headers: RequestHeaders;
+      progress: RequestProgress;
+      partialText?: string;
+    }
+  | {
+      type: 'upload-progress';
+      headers: RequestHeaders;
+      progress: RequestProgress;
+    }
+  | {
+      type: 'success';
+      headers: RequestHeaders;
+      response: Response;
+      expiresInTimestamp?: number;
+    }
+  | {
+      type: 'failure';
+      headers: RequestHeaders;
+      error: RequestError;
+    }
+  | {
+      type: 'cancel';
+      headers: RequestHeaders;
+    };